Remove boolean_query and response_query config fields.
The config would contain the existing query field which would specify a rule that returns either a boolean or an object value as the decision. Based on the decision type, response headers andbody can be added to the CheckResponse result. This would allow users to switch between boolean and non-boolean policy decisions by updating their policies and w/o re-configuing OPA.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>